+++ CHANGELOG.md
@@ -21,6 +21,8 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
 - Setting "replacespecial" is now by default set to "special"
 - Default mapping `<C-c>` will now copy in normal, pointer and visual mode as well
+- Setting "restoretabs" is now an enum that accepts different options
+- Setting "notificationforpermisions" is now an enum that accepts different options
 
 ### Removed